% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TravelTime2011.1dec}
\alias{TravelTime2011.1dec}
\title{Mean travel times to work, from 2011 ACS, rounded to 1 decimal place.}
\format{A data frame with 51 rows and 7 variables:
\describe{
  \item{Rank}{state rank, by estimated mean travel time, where 1 is lowest travel time and 51 is highest}
  \item{State}{full name of the state}
  \item{Estimate.1dec}{estimated mean travel time, in minutes}
  \item{MOE.1dec}{estimated Margin of Error (at the 90\% confidence level) of the estimated mean travel time, in minutes}
  \item{Abbreviation}{postal abbreviation of the state}
  \item{Region}{factor variable for geographic region of the state: Northeast, South, Midwest, West, Pacific}
  \item{FIPS}{Federal Information Processing Standard (FIPS) code of the state; may be useful for linking with other datasets}
}}
\source{
\url{https://www.census.gov/}
}
\usage{
TravelTime2011.1dec
}
\description{
A dataset containing the estimated mean travel time (in minutes) to work
of workers 16 years and over who did not work at home
(henceforth "mean travel time to work"), and its estimated Margin of Error
at the 90\% confidence level,
for each of the 51 states (including Washington, D.C.),
from the 2011 American Community Survey.
}
\details{
Due to rounding, some ranks are tied in this version of the data.
Also note that this dataset reports Margins of Error (MoEs)
instead of standard errors.
}
\keyword{datasets}
